Revert "[test] Repeat tests using cairo_push_group()."
This reverts commit 07122e64fa9529e7ba9323988a5af1d1e7c2c55f.

The extra testing did find a pdf bug, and that should be fixed,
but the extra maintenance burden of running another iteration
of all tests does not seem justfied at all---particularly since
it looks like dozens of new reference images would be needed
for the svg backend.

Also, the new "failures" of the image backend with this new
testing look like a misunderstanding of exactly what the new
testing is actually drawing.