Disable UniqueID font check in Type 1 subset font
Some Type 1 fonts have some PostScript code at the start of the font
that checks if the font is already cached in the printer (based on the
font UniqueID key). This breaks our subsetted font as it is not the
same as the original font.

Checking for and removing a random blob of PostScript is difficult so
this patch takes the simpler approach of searching for the PS code
that checks if the UniqueID dictionary entry exists

ie

  /UniqueID known

this code leaves the value "true" or "false" on the PS interpreter
stack depending on whether the "UniqueID" key exists.

Then the patch inserts the PS code " pop false " after the above code
to remove the old result and place "false" on the stack to make it
appear that UniqueID does not exist.

This approach should be able to handle any type of PostScript code
that attempts to do something with the UniqueID value in the font
dictionary.