PDF: Don't emit a new text object for every call to show_glyphs
A text object is a series of text operations enclosed in a 'BT'/'ET'
pair.

After a call to _cairo_pdf_operators_show_glyphs() the closing 'ET'
will not be emitted. This allows subsequent calls to show_glyphs() to
emit text into the same text object. A call to any other operator or
_cairo_pdf_operators_flush() will close the text object.