Bug: tessellator sometimes ends rightmost trapezoids too late
Reported on the cairo mailing list:

http://lists.cairographics.org/archives/cairo/2008-May/014233.html

The tessellator would sometimes produce self-intersecting trapezoids
because it would skip the last edge in the active list when deciding
whether we can continue the current trapezoid or not.

The bug never caused a problem with pixman based rasterisation
because pixman stops filling in a trapezoid once it detects a self
intersection.