[cairo-surface] Remove optional pattern argument from create_similar_solid()
This was added in 41c6eebcd1fab94fd3a91d09f1ef6ee0d8c7a044, to avoid
allocating short-lived patterns.  However, this was error prune as the
color information was duplicated in the pattern and could get out of
sync.  Indeed, it was out of sync before this commit in the call from
cairo-clip.c.

By allocating the solid pattern on the stack we fix the original problem
without creating new ones.