PDF: Don't use a Form dictionary for the content stream
because pdfTeX (and maybe other PDF consumers) can not handle it.

Previously the content stream was an Form dictionary because at the
time the content is written it is not known whether the content stream
will be belong to the page or be an XObject in a knockout group. With
the additional of the paginated surface function
set_fallback_images_required() this information is now known at the
time the content is emitted.