PS: execute type3 glyphs in their own dictionary
The PS emulation of the PDF text operators stores some data on the
user dict between calls to operators. When using the PDF text
operators in type 3 glyphs, the PDF operators were overwriting the
same data as the PDF operators in the page. Fix this by creating a new
dictionary to execute each type 3 glyph in.