Mark several new tests added during 1.7 as XFAIL.
The following tests were added during the 1.7 cycle to
demonstrate bugs:

	alpha-similar operator operator-alpha self-copy-overlap

In all cases, the identical bug existed in the last major
release, (1.6.4), so these are not new regressions, but merely
bugs that we're now aware of. The correct way to handle bugs
such as these is to mark them as expected failures (XFAIL).