[svg] Fix comp-op for paint, mask and show_glyphs operations.
In paint and show_glyphs, the compositing operator was not emitted at all.
In mask, the operator was also emitted for the mask itself, which is
wrong.

SVG clear and source differ from cairo as it also affects the
destination if the source pixel are completely transparent. We need to emit
an additional clip-to-self property.

Librsvg does not support clip-to-self, so it renders the SVG
test outputs incorrectly.

This patch also remove a lot of useless spaces in the style property
strings (I know, this should go in another commit).