[pattern] Fix acquiring outside of source bounds
Restrict the area we acquire from the source image to the bounds of
the source image, even when we have an identity matrix. This handles
circumstances where the pattern extents may be enlarged due to
filtering, for example when applying a bilinear filter.

This patch replaces the open-coded rectangle intersection which has
already proven itself to be a source of bugs.

(Patch ported to master by Chris Wilson, all bugs are his.)